+++ lib/montage.rb
@@ -1,6 +1,7 @@
-require 'yaml'
-require 'pathname'
 require 'digest/sha2'
+require 'erb'
+require 'pathname'
+require 'yaml'
 
 # Gems.
 require 'active_support/ordered_hash'
@@ -9,6 +10,7 @@ require 'rmagick'
 # On with the library...
 require 'montage/core_ext'
 require 'montage/project'
+require 'montage/sass_builder'
 require 'montage/source'
 require 'montage/sprite'
 require 'montage/version'